added profile photo for home page ---- added snooki pic when click --- cleaned routes
