+++ seeds/userData.json
@@ -7,7 +7,7 @@
     "biography": "Polizzi is from Marlboro, New York, and is an aspiring veterinary technician. She applied to an ad on Facebook that said 'Calling all guidos and guidettes,' which Polizzi said, 'was definitely about my lifestyle.' Polizzi earned the nickname 'Snooki' in middle school after being the first of her friends to 'make out' with a boy.",
     "date_of_birth": "1987-11-23",
     "sex": 1,
-    "profile_pic": "https://akns-images.eonline.com/eol_images/Entire_Site/2011427/300.ab.snooks.052711.jpg?fit=around%7C1200:1200&output-quality=90&crop=1200:1200;center,top",
+    "profile_pic": "https://ih1.redbubble.net/image.1244952539.6745/flat,128x128,075,f-pad,128x128,f8f8f8.jpg",
     "context": "The United States saw the election of Ronald Reagan as the 40th President, ushering in a period of conservative governance. The Summer Olympics were held in Moscow, although they were boycotted by many countries in protest against the Soviet invasion of Afghanistan. The eruption of Mount St. Helens in Washington State captured international attention, resulting in significant loss of life and widespread devastation. "
   },
   {
@@ -18,7 +18,7 @@
     "biography": "Giancola is from Hazlet, New Jersey and is recently single at the start of the series. She attended William Paterson University and was a midfielder on the women's soccer team. Giancola ultimately strikes up a relationship with fellow castmate Ortiz-Magro in season 1 episode 3.",
     "date_of_birth": "1987-03-14",
     "sex": 1,
-    "profile_pic": "https://static1.srcdn.com/wordpress/wp-content/uploads/2021/05/sammi-sweetheart-jersey-shore-cropped.jpg",
+    "profile_pic": "https://ih1.redbubble.net/image.1244952539.6745/flat,128x128,075,f-pad,128x128,f8f8f8.jpg",
     "context": "The United States saw the election of Ronald Reagan as the 40th President, ushering in a period of conservative governance. The Summer Olympics were held in Moscow, although they were boycotted by many countries in protest against the Soviet invasion of Afghanistan. The eruption of Mount St. Helens in Washington State captured international attention, resulting in significant loss of life and widespread devastation. "
   },
   {
@@ -76,4 +76,6 @@
     "profile_pic": "https://m.media-amazon.com/images/M/MV5BMTYzOTE2MjE1Ml5BMl5BanBnXkFtZTgwMDEwODg3MjE@._V1_.jpg",
     "context": "The United States saw the election of Ronald Reagan as the 40th President, ushering in a period of conservative governance. The Summer Olympics were held in Moscow, although they were boycotted by many countries in protest against the Soviet invasion of Afghanistan. The eruption of Mount St. Helens in Washington State captured international attention, resulting in significant loss of life and widespread devastation. "
   }
+ 
 ]
+
