+++ views/partials/card.hbs
@@ -4,7 +4,7 @@
       <div style="margin-right: 20px;">
         {{!-- Make the cloudinary links work --}}
         {{!-- <img src="{{ this.profile_pic }}">  --}}
-        <img src="../../public/assets/image/ProfilePhoto.png">
+        <img src={{ this.profile_pic }}>
       </div>
       <div style="font-weight: bold;">
         <h3>{{ this.first_name }} {{ this.last_name }}</h3>
