+++ views/partials/form_test.hbs
@@ -45,9 +45,9 @@
 </div>
 
 
-  <div class="form-group col-lg" id="drop-area">
+  {{!-- <div class="form-group col-lg" id="drop-area">
     <p>Drag and drop an image file here, or click to select.</p>
-  </div>
+  </div> --}}
 
   <div class="form-group col-lg">
   <label for="biography">Biography (max 100 words):</label>
