+++ views/partials/gen0.hbs
@@ -3,7 +3,7 @@
   <div data-user-id="{{ this.id }}"  class="card-container position-relative">
     <div class="avatar-container position-absolute" style="width: 60px;">
       <!-- Add code to place corresponding family member profile -->
-      <img width="50" height="50" class="avatar" src="https://i.dailymail.co.uk/i/pix/2013/03/01/article-2286814-183B611A000005DC-986_306x463.jpg" alt="">
+      <img width="50" height="50" class="avatar" src="https://i.pinimg.com/1200x/1c/51/90/1c5190da53e99bd9599eb44145ce1f47.jpg" alt="">
     </div>
     <a href="/api/user/card/{{this.id}}">
     <div class="name-container">{{ this.first_name }} <br> {{ this.last_name }}</div>
